+++ src/assets/scripts/client/AppController.js
@@ -82,6 +82,7 @@ export default class AppController {
      */
     setupHandlers() {
         this.onAirportChangeHandler = this.onAirportChange.bind(this);
+        this.onTrafficResetHandler = this.onTrafficReset.bind(this);
 
         return this;
     }
@@ -93,6 +94,7 @@ export default class AppController {
      */
     enable() {
         this._eventBus.on(EVENT.AIRPORT_CHANGE, this.onAirportChangeHandler);
+        this._eventBus.on(EVENT.TRAFFIC_RESET, this.onTrafficResetHandler);
 
         return this;
     }
@@ -104,6 +106,7 @@ export default class AppController {
      */
     disable() {
         this._eventBus.off(EVENT.AIRPORT_CHANGE, this.onAirportChangeHandler);
+        this._eventBus.off(EVENT.TRAFFIC_RESET, this.onTrafficResetHandler);
 
         return this.destroy();
     }
@@ -297,6 +300,21 @@ export default class AppController {
         SpawnScheduler.startScheduler();
 
         this.updateViewControls();
+    }
+
+    /**
+     * event callback fired from within the `TrafficRateController` when traffic should be cleared and respawned.
+     *
+     * @for AppController
+     * @method onTrafficReset
+     */
+    onTrafficReset() {
+        EventTracker.recordEvent(TRACKABLE_EVENT.AIRPORTS, 'traffic-reset', AirportController.current.icao);
+        this.aircraftController.aircraft_remove_all();
+        this.scopeModel.radarTargetCollection.reset();
+        AirportController.current.resetAllRunwayQueues();
+        SpawnScheduler.createPreSpawnDepartures();
+        SpawnScheduler.resetAirborneTraffic();
     }
 
     // TODO: this should live in a view class somewhere. temporary inclusion here to prevent tests from failing