+++ CHANGELOG.md
@@ -11,7 +11,7 @@
 - <a href="https://github.com/openscope/openscope/issues/1772" target="_blank">#1772</a> - Add/Update some Scandinavian airlines
 - <a href="https://github.com/openscope/openscope/issues/1878" target="_blank">#1878</a> - Update ASA/SWA airlines
 - <a href="https://github.com/openscope/openscope/issues/1878" target="_blank">#1878</a> - Add Embraer 175
-
+- <a href="https://github.com/openscope/openscope/issues/1345" target="_blank">#1345</a> - Add reset buttons to traffic config dialog
 
 # 6.26.0 (December 1, 2021)
 ### New Features